Move EntryProcessor#to_uri to Util and add spec
